Fix: initialise or destroy MyList programmeID's on successful login or logout.
